Remove legacy Starscream WebSocket and WebSocketProvider feature flag (#3)
The native URLSessionWebSocket implementation (added in 457a48d8) is now
proven in production. This removes the feature flag and all legacy code:

- Delete WebSocket.swift, WebSocketStream.swift, Compression.swift,
  SSLSecurity.swift, SSLClientCertificate.swift
- Relocate SOCKSProxyable, WSProtocol, and WebSocketConstants to
  URLSessionWebSocket.swift as top-level types
- Remove WebSocketProvider enum and simplify WebSocketTransport to
  always use URLSessionWebSocket
