+++ Sources/ApolloWebSocket/WebSocketTransport.swift
@@ -80,11 +80,6 @@ public class WebSocketTransport {
     }
   }
 
-  public enum WebSocketProvider {
-    case legacy
-    case urlSession
-  }
-
   public struct Configuration {
     /// The client name to use for this client. Defaults to `Self.defaultClientName`
     public fileprivate(set) var clientName: String
@@ -107,9 +102,6 @@ public class WebSocketTransport {
     /// The `OperationMessageIdCreator` used to generate a unique message identifier per request.
     /// Defaults to `ApolloSequencedOperationMessageIdCreator`.
     public let operationMessageIdCreator: any OperationMessageIdCreator
-    /// Which WebSocket implementation to use. Defaults to `.legacy`.
-    public let webSocketProvider: WebSocketProvider
-
     /// The designated initializer
     public init(
       clientName: String = WebSocketTransport.defaultClientName,
@@ -120,8 +112,7 @@ public class WebSocketTransport {
       connectOnInit: Bool = true,
       connectingPayload: JSONEncodableDictionary? = [:],
       requestBodyCreator: any RequestBodyCreator = ApolloRequestBodyCreator(),
-      operationMessageIdCreator: any OperationMessageIdCreator = ApolloSequencedOperationMessageIdCreator(),
-      webSocketProvider: WebSocketProvider = .legacy
+      operationMessageIdCreator: any OperationMessageIdCreator = ApolloSequencedOperationMessageIdCreator()
     ) {
       self.clientName = clientName
       self.clientVersion = clientVersion
@@ -132,7 +123,6 @@ public class WebSocketTransport {
       self.connectingPayload = connectingPayload
       self.requestBodyCreator = requestBodyCreator
       self.operationMessageIdCreator = operationMessageIdCreator
-      self.webSocketProvider = webSocketProvider
     }
   }
 
@@ -159,8 +149,7 @@ public class WebSocketTransport {
     }
   }
 
-  /// Convenience initializer that creates the appropriate `WebSocketClient` based on the
-  /// `webSocketProvider` in the configuration.
+  /// Convenience initializer that creates a `URLSessionWebSocket` for the given URL and protocol.
   ///
   /// - Parameters:
   ///   - url: The destination URL to connect to.
@@ -170,16 +159,11 @@ public class WebSocketTransport {
   ///             web socket connection. Defaults to a configuration with default values.
   public convenience init(
     url: URL,
-    webSocketProtocol: WebSocket.WSProtocol,
+    webSocketProtocol: WSProtocol,
     store: ApolloStore? = nil,
     config: Configuration = Configuration()
   ) {
-    let websocket: any WebSocketClient = switch config.webSocketProvider {
-    case .legacy:
-      WebSocket(url: url, protocol: webSocketProtocol)
-    case .urlSession:
-      URLSessionWebSocket(url: url, protocol: webSocketProtocol)
-    }
+    let websocket = URLSessionWebSocket(url: url, protocol: webSocketProtocol)
     self.init(websocket: websocket, store: store, config: config)
   }
 
@@ -463,14 +447,14 @@ public class WebSocketTransport {
 }
 
 extension URLRequest {
-  fileprivate var wsProtocol: WebSocket.WSProtocol? {
-    guard let header = value(forHTTPHeaderField: WebSocket.Constants.headerWSProtocolName) else {
+  fileprivate var wsProtocol: WSProtocol? {
+    guard let header = value(forHTTPHeaderField: WebSocketConstants.headerWSProtocolName) else {
       return nil
     }
 
     switch header {
-    case WebSocket.WSProtocol.graphql_transport_ws.description: return .graphql_transport_ws
-    case WebSocket.WSProtocol.graphql_ws.description: return .graphql_ws
+    case WSProtocol.graphql_transport_ws.description: return .graphql_transport_ws
+    case WSProtocol.graphql_ws.description: return .graphql_ws
     default: return nil
     }
   }
